APSActivator now supports 2 modes: Activator mode where it just scans all the bundles classes and manages any annotations. This is achieved by using the default constructor of APSActivator.
The second mode is the existing instances mode. This requires the constructor where you pass at least one instance to the constructor.  Calls to addManagedInstance(instance) can also be done before start(BundleContext) is done. In this case it will just add the provided instances as managed instances, and then scan the bundle classes skipping any class that already has a provided managed instance. This way web-container created instances like servlets can be managed by the activator also. In a WAB you mostly want to inject tracked services, but it is possible to provide service instances also.

APSVaadinOSGiApplication and APSOSGiSupport (for servlets, etc) creates an APSActivator and adds itself/its caller to it thus handling annotations.