Adapted to changes in APSConfigService admin API.
No longer uses the index to determine if a node for a "many" model is the first node representing the configuration itself
or an indexed instance of the configuraiton. The node selected event now passes on the "indexed" state of selected nodes.
So a config reference with a model returning true for "isMany" and indexed set to false means that the node is a top node
of an indexed configuration representing the configuration type, and will have a sub node for each index of the configuration
type. The sub nodes will have indexed set to true.

In the GUI it looks like this:

    datasources
        |
        +----datasource : 1  (represents 'datasource' as a value of 'datasources'. The '1' indicates that there are 1 instance).
                 |
                 +----datasource : 0  (represents instance 0 of datasource).

The first "datasource : 1" node will have indexed == false, while the second "datasource : 0" will have indexed == true.
ref.getEditValueModel().isMany() will be true for both.

So the GUI rendering of this no longer depends on the value of the index, which it never should have done from the
beginning!

Also new in the new admin API is that the edit model now only represents the configuration node, not an instance of it.
The new APSConfigReference represents a specific instance, and sits on both edit models (for full path) and index, and
configuration environment. It is the APSConfigReference which now generates the final key used to reference the config,
using the part keys from each edit model added to it, plus extra information like config env and index.

This resulted in that almost every place an APSConfigEditModel was passed a APSConfigReference is now passed instead.
When an APSConfigEditModel is actually needed it is gotten from the APSConfigReference.