Added feature to annotate private field with @ConfiguredInstance which specifies a config class (of APSConfigService type) and a '.' separated path to the instance name value in the configuration.
When this is available one instance per config entry found will be created and the name of each instance from the config will be injected into the instance, allowing it to pick the correct entry from the config for the instance.

This is the easiest and most flexible of all the supported ways of specifying multiple instances of a service when using APSActivator.

See the @ConfiguredInstance annotation javadoc for more info.