aps-9 General messaging dev activity
- CLM regeneration of source headers.
- Minor changes.

Left this in a non commited state a bit too long. Should always commit and push before longer breaks in working with the code.
I fucked up so badly that when I returned to this code there were uncommitted changes both on my MacBook Pro and on my iMac, and they were not the same! I have determined that the code on my iMac was much fresher than the code on my MacBook Pro which still contained some things I've thrown away. So I'm starting from the iMac code now. I'm pointing this out in a commit comment just to punish myself for the stupidity!