aps-6 Look over the REST API and handling.
Better and smarter now, but not annotation driven! @APSWebService is still available, but only for documentative purposes. Apache Karaf is so nasty that it does not include annotations on classes unless you at classes in your own bundle. bundleContext.getBundle().loadClass(...) does not provide annotation even if they are there and have retention RUNTIME. Thereby it wasn't possible to make the external protocol extender react on @APSWebService.

There are now 2 variants for doing REST calls using JSONREST as protocol. One is to provide service and method to call, and the other is to only provide the service and deduce the method based on the HTTP method used. When used that way each service would represent just one thing that can be created, read, updated, and deleted.

All services must be specified as externalizable either in bundle or in APSExternalizableService config.

Debugged / Tested.