I had "remove trailing spaces on ALL lines on save" checked in IDE config. This caused some trailing spaces in generated header to be stripped and then cause a lot of file changes when headers are regenerated. Thereby I change to "remove trailing spaces only on modified lines".
So this commit is a mix of newly added headers to files that has been creates since last header update and headers that have gone through trailing space cuts.

Should fix the header generator to not write spaces on spaces only lines. It is not #1 on my prio list however.