aps-168 Update the bus API & functionality on both backend and frontend. In the end I more or less kept the backend API, but with some cleanup. This due to different requirements on frontend and backend bus wise. The frontend bus API makes sense on the frontend, less so on the backend and on top of that the backend complicates such a solution due to OSGi and that APS don't require starting bundles in any specified order. The backend also has different messaging requirements than the frontend.
The messaging API on the backend no longer has a publish() and send() to
map to Vert.x eventbus. It needs to be more generic and handle any bus.
So now there is only send(), no publish(). The implementation wrapping
the Vert.x eventbus handles addresses starting with "all:" and if that
is the case the "all:" part is removed and publish() is done othewise
send() is done. APS uses the word "destination" rather than address,
and what a destination is depend on the implementation. The destination
is always a string.

Also fixed a workaround to a bug in APSActivator. The sender API had
an APSReplyableMessageSender that extended APSMessageSender. The Vert.x
eventbus implementation implemented APSReplyableMessageSender. APSActivator
could only see this API and not APSMessageSender so when a client with
type APSMessageSender was tracked no services were found. The annotation
supports specifying all implemented APIs, but still only the first is
found when looking for services. So as a workaround APSReplyableMessageSender
and APSMessageSender were merged into APSMessageSender, with the reply
API having a default implementation throwing an exception.